+++ src/treeView/repositories/repositories.ts
@@ -1,25 +1,3 @@
-// Actions should appear next to each item. Maybe it isnt possible
-
-// context item:
-
-// TODO: Find repositories (by dirs , with depth of 1.) Setting should have by default git.defaultCloneDirectory
-// TODO: globalState with added repo path that may be outside the dirs above.
-// TODO: on globalState paths check on startup, remove the key if repo isnt found.
-
-// If already cloned:
-// Open
-// Open in New Window
-// Add to Workspace
-// Remove from Local Files (?)
-
-// Not cloned:
-// Link to already cloned repo -> add to globalState
-// Clone [popup dest dir]
-// Clone to <Default Dir>
-
-// After cloning, show what to do with it dialogue (Open, new window, add to workspace, nothing(?))
-
-
 import vscode from 'vscode';
 import { BaseTreeDataProvider } from '../base';
 import { Repository, repositories } from '../../Repository/Repository';